Sauvegarde et chargement de partie + bugfix de l'affichage du mode de jeu en titre de la fenêtre

+++ lib/HudModeDeJeu.rb
@@ -1,5 +1,5 @@
-# Instance du menu de sélection des modes de jeu : permet la sélection 
-# entre l'un des quatre modes, le chargement d'une sauvegarde, l'accès au profil, 
+# Instance du menu de sélection des modes de jeu : permet la sélection
+# entre l'un des quatre modes, le chargement d'une sauvegarde, l'accès au profil,
 # aux options, et un retour au menu de connexion
 
 # TODO : bouton d'accès au menu de connexion
@@ -27,6 +27,7 @@ class HudModeDeJeu < Hud
 		# TODO - foutus nombres magiques
 
 		self.attach(@btnSauvegarde,varX,varY-2,2,1)
+		self.attach(@lblDescription, varX, varY-1, 2, 1)
 
 		self.attach(@btnTutoriel,varX, varY+1, 2, 1)
 
@@ -35,7 +36,7 @@ class HudModeDeJeu < Hud
 		self.attach(@btnChrono,varX, varY+6, 2, 3)
 
 		self.attach(@btnExplo,varX, varY+9, 2, 3)
-		
+
 		self.attach(@btnOptions, 1, varY+14, 1, 1)
 		self.attach(@btnQuitter, varX+4, varY+14, 1, 1)
 		self.attach(@btnProfil, varX+4, varY-4, 1, 1)
@@ -47,18 +48,24 @@ class HudModeDeJeu < Hud
 	# Return self
 	# TODO : gérer l'exception ERRNOENT si pas de fichier (afficher un popup)
 	def initBoutonChargerSauvegarde
-		@btnSauvegarde = Gtk::Button.new :label => "Charger une sauvegarde"
+		@btnSauvegarde = Gtk::Button.new :label => "Charger la dernière sauvegarde"
 		@btnSauvegarde.signal_connect('clicked') do
-			File.open("saves/"+@@name+".txt", 'r') do |f|
-				dataLoaded=Marshal.load(f)
-				grille=dataLoaded[0]
-				@@mode=dataLoaded[1]
-				@difficulte=dataLoaded[2]
-				case @@mode
-					when :explo then lancementExplo(grille)
-					when :rapide then lancementRapide(grille)
-					when :tutoriel then lancementTutoriel(grille)
-					when :aventure then lancementAventure(grille)
+			if !Dir.exist?("saves")
+				self.setDesc("Le dossier de sauvegarde n'existe pas !")
+			elsif !File.exist?("saves/"+@@name+".txt")
+				self.setDesc("Le fichier de sauvegarde \"" + @@name + "\" n'existe pas !")
+			else
+				File.open("saves/"+@@name+".txt", 'r') do |f|
+					dataLoaded=Marshal.load(f)
+					grille=dataLoaded[0]
+					@@mode=dataLoaded[1]
+					@@difficulte=dataLoaded[2]
+					case @@mode
+						when :explo then lancementExplo(grille)
+						when :rapide then lancementRapide(grille)
+						when :tutoriel then lancementTutoriel(grille)
+						when :aventure then lancementAventure(grille)
+					end
 				end
 			end
 		end
@@ -82,7 +89,7 @@ class HudModeDeJeu < Hud
 		end
 		self
 	end
-	
+
 	# Crée et connecte le bouton de lancement du mode explo
 	# Return self
 	def initBoutonExplo
@@ -104,7 +111,7 @@ class HudModeDeJeu < Hud
 		end
 		self
 	end
-	
-	protected 
+
+	protected
 		attr_reader :btnTutoriel, :btnExploFacile, :btnExploMoy
 end
